% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fontionsChisquared.R
\name{chi_deux}
\alias{chi_deux}
\title{Test du khi-deux pour les tableaux de contingence et les vecteurs de fréquences}
\usage{
chi_deux(
  x,
  y = NULL,
  correct = TRUE,
  p = rep(1/length(x), length(x)),
  rescale.p = FALSE,
  simulate.p.value = FALSE,
  B = 2000
)
}
\arguments{
\item{x}{Un tableau de contingence ou un vecteur de fréquences observées.}

\item{y}{Un vecteur de fréquences observées (facultatif).}

\item{correct}{Logique, si TRUE, applique la correction de continuité de Yates pour les tableaux 2x2.}

\item{p}{Un vecteur de probabilités théoriques pour les fréquences observées.}

\item{rescale.p}{Logique, si TRUE, les probabilités sont mises à l'échelle pour totaliser 1.}

\item{simulate.p.value}{Logique, si TRUE, utilise une simulation pour calculer la valeur p.}

\item{B}{Nombre de réplicats pour la simulation de la valeur p.}
}
\value{
Un objet de classe "htest" contenant les résultats du test du khi-deux.
}
\description{
Cette fonction effectue un test du khi-deux pour les tableaux de contingence et les vecteurs de fréquences.
}
\examples{
# Exemple avec un tableau de contingence
dataE <- matrix(c(50, 100, 120, 80, 130, 20), nrow = 3, byrow = TRUE,
               dimnames = list(c("Primaire", "Secondaire", "Supérieur"), c("Employé", "Chômeur")))
chi_deux(dataE)

# Exemple avec des vecteurs de fréquences observées et théoriques
obs <- c(50, 100, 120)
theo <- c(0.2, 0.3, 0.5)
chi_deux(obs, p = theo)
}
